Refactor InvalidLogsHelper for improved logging and module structure
